+++ packages/s2-core/__tests__/unit/interaction/selected-cell-move-spec.ts
@@ -71,6 +71,9 @@ describe('Interaction Keyboard Move Tests', () => {
   test('should move selected cell right', () => {
     s2.interaction.changeState = jest.fn((state) => {});
     s2.interaction.getCells = () => [mockCell00.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell00.mockCell;
+    keyboardMove.endCell = mockCell00.mockCell;
 
     s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
       key: InteractionKeyboardKey.ARROW_RIGHT,
@@ -92,6 +95,9 @@ describe('Interaction Keyboard Move Tests', () => {
   test('should move selected cell left', () => {
     s2.interaction.changeState = jest.fn((state) => {});
     s2.interaction.getCells = () => [mockCell01.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell01.mockCell;
+    keyboardMove.endCell = mockCell01.mockCell;
 
     s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
       key: InteractionKeyboardKey.ARROW_LEFT,
@@ -114,6 +120,9 @@ describe('Interaction Keyboard Move Tests', () => {
   test('should move selected cell up', () => {
     s2.interaction.changeState = jest.fn((state) => {});
     s2.interaction.getCells = () => [mockCell10.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell10.mockCell;
+    keyboardMove.endCell = mockCell10.mockCell;
 
     s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
       key: InteractionKeyboardKey.ARROW_UP,
@@ -136,6 +145,9 @@ describe('Interaction Keyboard Move Tests', () => {
   test('should move selected cell down', () => {
     s2.interaction.changeState = jest.fn((state) => {});
     s2.interaction.getCells = () => [mockCell01.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell01.mockCell;
+    keyboardMove.endCell = mockCell01.mockCell;
 
     s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
       key: InteractionKeyboardKey.ARROW_DOWN,
@@ -153,5 +165,126 @@ describe('Interaction Keyboard Move Tests', () => {
       key: InteractionKeyboardKey.ARROW_DOWN,
     } as KeyboardEvent);
     expect(s2.interaction.changeState).not.toBeCalled();
+  });
+
+  test('should move selected with meta', () => {
+    s2.interaction.changeState = jest.fn((state) => {});
+    s2.interaction.getCells = () => [mockCell00.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell00.mockCell;
+    keyboardMove.endCell = mockCell00.mockCell;
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_RIGHT,
+      metaKey: true,
+    } as KeyboardEvent);
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [{ colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' }],
+      stateName: 'selected',
+    });
+
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_DOWN,
+      metaKey: true,
+    } as KeyboardEvent);
+
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [{ colIndex: 1, id: '1-1', rowIndex: 1, type: 'dataCell' }],
+      stateName: 'selected',
+    });
+
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_LEFT,
+      metaKey: true,
+    } as KeyboardEvent);
+
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [{ colIndex: 0, id: '1-0', rowIndex: 1, type: 'dataCell' }],
+      stateName: 'selected',
+    });
+
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_UP,
+      metaKey: true,
+    } as KeyboardEvent);
+
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [{ colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' }],
+      stateName: 'selected',
+    });
+  });
+
+  test('should move selected with shift', () => {
+    s2.interaction.changeState = jest.fn((state) => {});
+    s2.interaction.getCells = () => [mockCell00.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell00.mockCell;
+    keyboardMove.endCell = mockCell00.mockCell;
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_RIGHT,
+      shiftKey: true,
+    } as KeyboardEvent);
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [
+        { colIndex: 0, id: '0-0', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' },
+      ],
+      stateName: 'selected',
+    });
+
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_DOWN,
+      shiftKey: true,
+    } as KeyboardEvent);
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [
+        { colIndex: 0, id: '0-0', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 0, id: '1-0', rowIndex: 1, type: 'dataCell' },
+        { colIndex: 1, id: '1-1', rowIndex: 1, type: 'dataCell' },
+      ],
+      stateName: 'selected',
+    });
+  });
+
+  test('should move selected with shift and meta', () => {
+    s2.interaction.changeState = jest.fn((state) => {});
+    s2.interaction.getCells = () => [mockCell00.mockCell as any];
+    // select cell
+    keyboardMove.startCell = mockCell00.mockCell;
+    keyboardMove.endCell = mockCell00.mockCell;
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_RIGHT,
+      shiftKey: true,
+      metaKey: true,
+    } as KeyboardEvent);
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [
+        { colIndex: 0, id: '0-0', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' },
+      ],
+      stateName: 'selected',
+    });
+
+    s2.emit(S2Event.GLOBAL_KEYBOARD_DOWN, {
+      key: InteractionKeyboardKey.ARROW_DOWN,
+      shiftKey: true,
+    } as KeyboardEvent);
+    expect(s2.interaction.changeState).toBeCalled();
+    expect(s2.interaction.changeState).toBeCalledWith({
+      cells: [
+        { colIndex: 0, id: '0-0', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 1, id: '0-1', rowIndex: 0, type: 'dataCell' },
+        { colIndex: 0, id: '1-0', rowIndex: 1, type: 'dataCell' },
+        { colIndex: 1, id: '1-1', rowIndex: 1, type: 'dataCell' },
+      ],
+      stateName: 'selected',
+    });
   });
 });