+++ packages/s2-core/__tests__/unit/sheet-type/pivot-sheet-spec.ts
@@ -2,7 +2,7 @@
 import { getContainer } from 'tests/util/helpers';
 import * as dataCfg from 'tests/data/simple-data.json';
 import { Canvas, Event as GEvent } from '@antv/g-canvas';
-import { cloneDeep } from 'lodash';
+import { cloneDeep, get, last } from 'lodash';
 import { PivotSheet, SpreadSheet } from '@/sheet-type';
 import {
   CellTypes,
@@ -20,6 +20,7 @@ import {
 import { Node } from '@/facet/layout/node';
 import { customMerge, getSafetyDataConfig } from '@/utils';
 import { BaseTooltip } from '@/ui/tooltip';
+import { CornerCell } from '@/cell/corner-cell';
 
 const originalDataCfg = cloneDeep(dataCfg);
 
@@ -493,16 +494,16 @@ describe('PivotSheet Tests', () => {
     expect(s2.getColumnNodes()).toHaveLength(3);
   });
 
-  test('should change sheet size', () => {
+  test('should change sheet container size', () => {
     s2.changeSheetSize(1000, 500);
 
-    expect(s2.options.width).toEqual(1000);
-    expect(s2.options.height).toEqual(500);
+    expect(s2.options.width).toStrictEqual(1000);
+    expect(s2.options.height).toStrictEqual(500);
 
     const canvas = s2.container.get('el') as HTMLCanvasElement;
 
-    expect(canvas.style.width).toEqual(`1000px`);
-    expect(canvas.style.height).toEqual(`500px`);
+    expect(canvas.style.width).toStrictEqual(`1000px`);
+    expect(canvas.style.height).toStrictEqual(`500px`);
   });
 
   test('should set display:block style with canvas', () => {
@@ -534,6 +535,46 @@ describe('PivotSheet Tests', () => {
     expect(s2.panelGroup.getChildren()).toHaveLength(1);
     expect(s2.panelGroup.findAllByName(KEY_GROUP_PANEL_SCROLL)).toHaveLength(1);
   });
+
+  test.each([
+    {
+      width: s2Options.width + 100,
+      height: s2Options.height + 100,
+    },
+    {
+      width: s2Options.width + 100,
+      height: s2Options.height,
+    },
+    {
+      width: s2Options.width,
+      height: s2Options.height + 100,
+    },
+    {
+      width: s2Options.width,
+      height: s2Options.height,
+    },
+  ])(
+    'should skip change sheet container size if width and height not changed %o',
+    ({ width, height }) => {
+      s2.changeSheetSize(s2Options.width, s2Options.height);
+
+      const isCalled = width !== s2Options.width || height !== s2Options.height;
+
+      const changeSizeSpy = jest
+        .spyOn(s2.container, 'changeSize')
+        .mockImplementationOnce(() => {});
+
+      s2.changeSheetSize(width, height);
+
+      expect(s2.options.width).toStrictEqual(
+        isCalled ? width : s2.options.width,
+      );
+      expect(s2.options.height).toStrictEqual(
+        isCalled ? height : s2.options.height,
+      );
+      expect(changeSizeSpy).toHaveBeenCalledTimes(isCalled ? 1 : 0);
+    },
+  );
 
   test('should init column nodes', () => {
     // [type -> cost, type -> price] => [笔 -> cost, 笔 -> price]
@@ -579,6 +620,48 @@ describe('PivotSheet Tests', () => {
     ).toBeFalsy();
 
     renderSpy.mockRestore();
+  });
+
+  test('should get extra field text', () => {
+    const pivotSheet = new PivotSheet(
+      container,
+      customMerge(originalDataCfg, {
+        fields: {
+          valueInCols: false,
+        },
+      }),
+      s2Options,
+    );
+    pivotSheet.render();
+
+    const extraField = last(
+      pivotSheet.facet.cornerHeader.getChildren(),
+    ) as CornerCell;
+    expect(get(extraField, 'actualText')).toEqual('数值');
+  });
+
+  // https://github.com/antvis/S2/issues/1212
+  test('should get custom extra field text', () => {
+    const cornerExtraFieldText = 'custom';
+
+    const pivotSheet = new PivotSheet(
+      container,
+      customMerge(originalDataCfg, {
+        fields: {
+          valueInCols: false,
+        },
+      }),
+      {
+        ...s2Options,
+        cornerExtraFieldText,
+      },
+    );
+    pivotSheet.render();
+
+    const extraField = last(
+      pivotSheet.facet.cornerHeader.getChildren(),
+    ) as CornerCell;
+    expect(get(extraField, 'actualText')).toEqual(cornerExtraFieldText);
   });
 
   describe('Tree Collapse Tests', () => {