+++ packages/s2-core/__tests__/unit/utils/facet-spec.ts
@@ -34,4 +34,20 @@ describe('Facet util test', () => {
       end: 7,
     });
   });
+
+  test('should get correct index range for invalid input', () => {
+    const offsets = [0, 30, 60, 90, 120, 150, 160, 170, 190];
+    expect(getIndexRangeWithOffsets(offsets, 0, -10)).toStrictEqual({
+      start: 0,
+      end: 0,
+    });
+  });
+
+  test('should get correct index range with equal min and max height', () => {
+    const offsets = [0, 30, 60, 90, 120, 150, 160, 170, 190];
+    expect(getIndexRangeWithOffsets(offsets, 60, 60)).toStrictEqual({
+      start: 2,
+      end: 2,
+    });
+  });
 });