+++ packages/s2-core/src/cell/corner-cell.ts
@@ -24,7 +24,11 @@ import {
   ResizeDirectionType,
   S2Event,
 } from '@/common/constant';
-import { CellBorderPosition, TextTheme } from '@/common/interface';
+import {
+  CellBorderPosition,
+  FormatResult,
+  TextTheme,
+} from '@/common/interface';
 import { CornerHeaderConfig } from '@/facet/header/corner';
 import {
   getTextPosition,
@@ -67,21 +71,18 @@ export class CornerCell extends HeaderCell {
   }
 
   protected drawCellText() {
-    const { label } = this.meta;
-
     const { x } = this.getContentArea();
     const { y, height } = this.getCellArea();
 
     const textStyle = this.getTextStyle();
-
-    const cornerText = this.getCornerText(label);
+    const cornerText = this.getCornerText();
 
     // 当为树状结构下需要计算文本前收起展开的icon占的位置
 
     const maxWidth = this.getMaxTextWidth();
     const text = getEllipsisText({
       text: cornerText,
-      maxWidth: maxWidth,
+      maxWidth,
       fontParam: textStyle,
       placeholder: this.spreadsheet.options.placeholder,
     });
@@ -100,7 +101,7 @@ export class CornerCell extends HeaderCell {
       // 第二行重新计算...逻辑
       secondLine = getEllipsisText({
         text: secondLine,
-        maxWidth: maxWidth,
+        maxWidth,
         fontParam: textStyle,
       });
     }
@@ -108,9 +109,9 @@ export class CornerCell extends HeaderCell {
     const { x: textX } = getTextPosition(
       {
         x: x + this.getTreeIconWidth(),
-        y: y,
+        y,
         width: maxWidth,
-        height: height,
+        height,
       },
       textStyle,
     );
@@ -355,8 +356,27 @@ export class CornerCell extends HeaderCell {
     };
   }
 
-  protected getCornerText(label: string): string {
-    if (isEqual(label, EXTRA_FIELD)) {
+  // corner cell 不需要使用formatter进行格式化
+  protected getFormattedFieldValue(): FormatResult {
+    const { label, field } = this.meta;
+
+    if (!isEqual(field, EXTRA_FIELD)) {
+      return {
+        formattedValue: label,
+        value: label,
+      };
+    }
+
+    const fieldName = this.spreadsheet.dataSet.getFieldName(label);
+
+    return {
+      formattedValue: fieldName || label,
+      value: label,
+    };
+  }
+
+  protected getCornerText(): string {
+    if (isEqual(this.meta.label, EXTRA_FIELD)) {
       return this.spreadsheet.options?.cornerText || DEFAULT_CORNER_TEXT;
     }