+++ packages/s2-core/src/interaction/data-cell-multi-selection.ts
@@ -1,11 +1,13 @@
 import { Event } from '@antv/g-canvas';
 import { isEmpty } from 'lodash';
-import { getCellMeta } from 'src/utils/interaction/select-event';
+import {
+  getCellMeta,
+  isMultiSelectionKey,
+} from 'src/utils/interaction/select-event';
 import { BaseEvent, BaseEventImplement } from './base-interaction';
 import { getActiveCellsTooltipData } from '@/utils/tooltip';
 import {
   InterceptType,
-  InteractionKeyboardKey,
   InteractionStateName,
   S2Event,
 } from '@/common/constant';
@@ -28,7 +30,7 @@ export class DataCellMultiSelection
     this.spreadsheet.on(
       S2Event.GLOBAL_KEYBOARD_DOWN,
       (event: KeyboardEvent) => {
-        if (event.key === InteractionKeyboardKey.META) {
+        if (isMultiSelectionKey(event)) {
           this.isMultiSelection = true;
           this.spreadsheet.interaction.addIntercepts([InterceptType.CLICK]);
         }
@@ -38,7 +40,7 @@ export class DataCellMultiSelection
 
   private bindKeyboardUp() {
     this.spreadsheet.on(S2Event.GLOBAL_KEYBOARD_UP, (event: KeyboardEvent) => {
-      if (event.key === InteractionKeyboardKey.META) {
+      if (isMultiSelectionKey(event)) {
         this.isMultiSelection = false;
         this.spreadsheet.interaction.removeIntercepts([InterceptType.CLICK]);
       }