+++ packages/s2-core/src/utils/export/copy.ts
@@ -30,10 +30,10 @@ const getFiledIdFromMeta = (meta: CellMeta, spreadsheet: SpreadSheet) => {
 };
 
 const getHeaderNodeFromMeta = (meta: CellMeta, spreadsheet: SpreadSheet) => {
-  const [rowId, colId] = meta.id.split('-');
+  const { rowIndex, colIndex } = meta;
   return [
-    spreadsheet.getRowNodes().find((row) => row.id === rowId),
-    spreadsheet.getColumnNodes().find((col) => col.id === colId),
+    spreadsheet.getRowNodes().find((row) => row.rowIndex === rowIndex),
+    spreadsheet.getColumnNodes().find((col) => col.colIndex === colIndex),
   ];
 };
 
@@ -58,7 +58,7 @@ const getValueFromMeta = (
         ...rowNode.query,
         ...colNode.query,
       },
-      rowNode: rowNode,
+      rowNode,
     });
     return cell[VALUE_FIELD];
   }
@@ -77,8 +77,8 @@ const format = (
 
 export const convertString = (v: string) => {
   if (/\n/.test(v)) {
-    // 单元格内换行
-    return '"' + v.replace(/\r\n?/g, '\n') + '"';
+    // 单元格内换行 替换双引号 防止内容存在双引号 导致内容换行出错
+    return '"' + v.replace(/\r\n?/g, '\n').replace(/"/g, "'") + '"';
   }
   return v;
 };