+++ packages/s2-react/__tests__/unit/utils/drill-down-spec.ts
@@ -104,7 +104,7 @@ describe('Drill Down Test', () => {
     expect(mockInstance.store.get('drillDownIdPathMap')).not.toBeEmpty();
   });
 
-  test('for handleDrillDownIcon function', async () => {
+  test('for handleDrillDownIcon function', () => {
     const mergedOptions = handleDrillDownIcon(
       {
         options: mockInstance.options,
@@ -133,7 +133,7 @@ describe('Drill Down Test', () => {
     );
   });
 
-  test('for getDrillDownCache function', async () => {
+  test('for getDrillDownCache function', () => {
     const mockDrillDownDataCache = {
       rowId: 'root[&]浙江省[&]杭州市',
       drillLevel: 0,