+++ s2-site/examples/custom/custom-cell/demo/corner-header.ts
@@ -4,10 +4,10 @@ import { get } from 'lodash';
 
 // 自定义角头单元格，实现特有功能
 class CustomCornerHeader extends Group {
-  protected node;
-  protected backgroundShape;
-  protected textShape;
-  public constructor(node) {
+  node;
+  backgroundShape;
+  textShape;
+  constructor(node) {
     super({});
     this.node = node;
     this.initCornerHeader();
@@ -65,6 +65,9 @@ fetch(
     const s2Options = {
       width: 600,
       height: 480,
+      interaction: {
+        hoverHighlight: false,
+      },
       cornerHeader: (node, s2, headConfig) => {
         return new CustomCornerHeader(node);
       },